requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.38
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.3.0
networkx==3.4.2
numpy==2.2.5
openai==1.76.2
orjson==3.10.18
packaging==24.2
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.25.5
PyMySQL==1.1.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==75.8.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wheel==0.45.1
zstandard==0.23.0
